pyproject.toml
[tool.poetry]
name = "portfolio-manager-adk"
version = "0.1.0"
description = "Backend-only AI agent system for portfolio management using Google ADK."
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "agent" },
    { include = "data" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-aiplatform = "^1.38.1"
google-adk = "^0.1.0"
textblob = "^0.17.1"
pandas = "^2.1.4"
numpy = "^1.26.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
finnhub-python = "^2.4.18"
deprecated = "^1.2.18"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
